fix: DungTT migration

+++ database/migrations/2023_05_02_083718_create_genre_manga_table.php
@@ -16,7 +16,7 @@ class CreateGenreMangaTable extends Migration
         Schema::create('genre_manga', function (Blueprint $table) {
             $table->id();
             $table->tinyInteger('genre_id');
-            $table->tinyInteger('manga_id');
+            $table->integer('manga_id');
         });
     }
